# app/docs/swagger.yaml
openapi: 3.0.2
info:
  title: Bookman API
  version: '1.0'
  description: Intelligent book management system with Llama3 integration.
servers:
  - url: http://localhost:5000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        year_published:
          type: integer
        summary:
          type: string

    Review:
      type: object
      properties:
        id:
          type: integer
        book_id:
          type: integer
        user_id:
          type: integer
        review_text:
          type: string
        rating:
          type: integer

paths:
  /books/:
    get:
      summary: Get all books
      responses:
        '200':
          description: List of books
    post:
      summary: Add a new book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Book created

  /books/{book_id}:
    get:
      summary: Retrieve a book by ID
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book found
    put:
      summary: Update a book by ID
      security:
        - bearerAuth: []
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated
    delete:
      summary: Delete a book by ID
      security:
        - bearerAuth: []
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Book deleted

  /books/{book_id}/reviews:
    get:
      summary: Get all reviews for a book
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of reviews
    post:
      summary: Add a review for a book
#      security:
#        - bearerAuth: []
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '201':
          description: Review created

  /books/{book_id}/summary:
    get:
      summary: Get summary and aggregated rating for a book
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Summary and rating

  /generate-summary:
    post:
      summary: Generate a summary using Llama3
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                book_id:
                  type: integer
                text:
                  type: string
                method:
                  type: string
              required:
                - book_id
                - text
      responses:
        '201':
          description: Summary generated

  /recommendations:
    get:
      summary: Get book recommendations based on user preferences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of recommended books
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: akash
                email:
                  type: string
                  example: akash@example.com
                password:
                  type: string
                  example: secure123
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input or user already exists

  /api/auth/login:
    post:
      summary: Login and receive JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: akash@example.com
                password:
                  type: string
                  example: secure123
              required:
                - email
                - password
      responses:
        '200':
          description: JWT token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJ0eXAiOiJKV1QiLCJhbGciOi...
        '401':
          description: Invalid credentials